refactor: rename trigger-manager.ts to trigger_manager.ts
